+++ sdk/nodejs/tests/cloudwatch/index.ts
@@ -1,0 +1,1 @@
+import * as cloudwatchTests from "./cloudwatch.spec";
