+++ sdk/nodejs/tests/index.ts
@@ -1,0 +1,1 @@
+import * as cloudwatch from "./cloudwatch";
