+++ sdk/nodejs/tests/package.json
@@ -18,15 +18,16 @@
         "builtin-modules": "3.0.0",
         "mime": "^2.0.0",
         "read-package-tree": "^5.2.1",
-        "resolve": "^1.7.1"
+        "resolve": "^1.7.1",
+        "mocha": "^3.5.0",
+        "typescript": "^3.0.0",
+        "chai": "^4.2.0",
+        "proxyquire": "^2.1.0"
     },
     "devDependencies": {
         "@types/mime": "^2.0.0",
         "@types/mocha": "^2.2.42",
-        "@types/node": "^10.0.0",
-        "istanbul": "^0.4.5",
-        "mocha": "^3.5.0",
-        "typescript": "^3.0.0"
+        "@types/node": "^10.0.0"
     },
     "pulumi": {
         "resource": true
